src/bot/requirements.txt
botbuilder-core
botbuilder-schema
aiohttp
requests
python-dotenv